+++ package.json
@@ -24,6 +24,7 @@
     "wagmi": "^2.12.30"
   },
   "devDependencies": {
+    "@iconify/react": "^5.0.2",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
