--- src/app/(landing)/types.ts
@@ -1,4 +1,0 @@
-export interface PawnItem {
-  item: string;
-  usdtValue: number;
-}
