+++ src/app/pawn/enum.ts
@@ -1,0 +1,7 @@
+export const PAWN = 'PAWN';
+export const REDEEM = 'REDEEM';
+
+export const wei = 1000000000000000000;
+export const gwei = 1000000000;
+
+export const PAWN_PRICE = 100;
