+++ src/app/pawn/types.ts
@@ -1,0 +1,4 @@
+export interface PawnItem {
+  item: string;
+  usdtValue: number;
+}
