+++ src/components/Button/index.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useCallback, useMemo } from 'react';
 import {
   BUTTON_SIZES,
