+++ src/components/Text/enum.ts
@@ -6,34 +6,30 @@ export interface TextVariants {
 export const TEXT_VARIANTS = {
   HEADING1: {
     element: 'h1' as keyof JSX.IntrinsicElements,
-    className: 'laptop:text-h1',
+    className: 'font-roboto font-black laptop:text-h1',
   },
   HEADING2: {
     element: 'h2' as keyof JSX.IntrinsicElements,
-    className: 'laptop:text-h2',
+    className: 'font-roboto font-black laptop:text-h2',
   },
   HEADING3: {
     element: 'h3' as keyof JSX.IntrinsicElements,
-    className: 'laptop:text-h3',
+    className: 'font-roboto font-black laptop:text-h3',
   },
   HEADING4: {
     element: 'h4' as keyof JSX.IntrinsicElements,
-    className: 'laptop:text-h4',
+    className: 'font-roboto font-black laptop:text-h4',
   },
   HEADING5: {
     element: 'h5' as keyof JSX.IntrinsicElements,
-    className: 'laptop:text-h5',
-  },
-  HEADING6: {
-    element: 'h6' as keyof JSX.IntrinsicElements,
-    className: 'laptop:text-h6',
+    className: 'font-roboto font-black laptop:text-h5',
   },
   BODY: {
     element: 'p' as keyof JSX.IntrinsicElements,
-    className: 'laptop:text-body',
+    className: 'font-light laptop:text-body',
   },
   CAPTION: {
     element: 'span' as keyof JSX.IntrinsicElements,
-    className: 'laptop:text-caption',
+    className: 'font-light laptop:text-caption',
   },
 };
